🔨 refactor(interfaceBuilder.ts): improve interface generation logic and add support for ContentTypes and ContentType 📝 chore(interfaceBuilder.ts): update Payload interface to use ContentType and add comments for MediaFormat interface The interface generation logic in `interfaceBuilder.ts` has been improved to generate the `ContentTypes` interface and the `ContentType` type. The `ContentTypes` interface now includes the `User`, `Media`, and `MediaFormat` interfaces in addition to the interfaces generated from the schemas. The `ContentType` type is a generic type that allows accessing the specific interface based on the key of `ContentTypes`.
The `Payload` interface has been updated to use the `ContentType` type instead of a generic type. The `data` property of the `Payload` interface now expects an array of `ContentType<T>` where `T` is a key of `ContentTypes`. Additionally, comments have been added to the `MediaFormat` interface to provide a description of each property.