🔧 chore(README.md): update config object in the README to include new property contentTypesToIgnore 📝 docs(README.md): update description for alwaysAddEnumSuffix property in the config object
The config object in the README has been updated to include a new property called contentTypesToIgnore. This property is an array that contains the UIDs of the content types that should be ignored by the plugin. Additionally, the description for the alwaysAddEnumSuffix property has been updated to clarify that when set to true, the generated enums will have a suffix that is the content type name. For example, if the content type is Car, the enum would be named CarType_Car. These changes improve the documentation and provide more clarity on the configuration options available.